+++ lib/ansible/ruby/dsl_builders/tasks_spec.rb
@@ -47,6 +47,90 @@ describe Ansible::Ruby::DslBuilders::Tasks do
     end
   end
 
+  context 'include' do
+    subject(:inclusion) { tasks.items.find { |item| item.is_a?(Ansible::Ruby::Models::Inclusion) } }
+
+    context 'with tasks' do
+      subject { evaluate }
+
+      let(:ruby) do
+        <<-RUBY
+        ansible_include '/some_file.yml'
+
+        task 'Copy something' do
+            copy do
+              src '/file1.conf'
+              dest '/file2.conf'
+            end
+        end
+        RUBY
+      end
+
+      it { is_expected.to be_a Ansible::Ruby::Models::Tasks }
+      it do
+        is_expected.to have_attributes items: include(be_a(Ansible::Ruby::Models::Task),
+                                                      be_a(Ansible::Ruby::Models::Inclusion))
+      end
+
+      describe 'inclusion' do
+        subject { inclusion }
+
+        it { is_expected.to have_attributes file: '/some_file.yml' }
+      end
+    end
+
+    context 'with variables' do
+      let(:ruby) do
+        <<-RUBY
+          ansible_include '/some_file.yml' do
+            static true
+            variables stuff: true
+          end
+        RUBY
+      end
+
+      it { is_expected.to be_a Ansible::Ruby::Models::Inclusion }
+      it do
+        is_expected.to have_attributes file: '/some_file.yml',
+                                       static: true,
+                                       variables: { stuff: true }
+      end
+    end
+
+    context 'jinja' do
+      let(:ruby) do
+        <<-RUBY
+        ansible_include '/some_file.yml' do
+          static true
+          variables stuff: jinja('toodles')
+        end
+        RUBY
+      end
+
+      it { is_expected.to be_a Ansible::Ruby::Models::Inclusion }
+      it do
+        is_expected.to have_attributes file: '/some_file.yml',
+                                       variables: { stuff: '{{ toodles }}' }
+      end
+    end
+
+    context 'static' do
+      let(:ruby) do
+        <<-RUBY
+        ansible_include '/some_file.yml' do
+          static true
+        end
+        RUBY
+      end
+
+      it { is_expected.to be_a Ansible::Ruby::Models::Inclusion }
+      it do
+        is_expected.to have_attributes file: '/some_file.yml',
+                                       static: true
+      end
+    end
+  end
+
   context 'invalid method' do
     let(:ruby) { 'foobar()' }
     subject { -> { evaluate } }